Move common.js out of node_modules to prevent it from being deleted when `npm install`.
