+++ examples/common.js
